requirements.txt
llama-cpp-python
numpy
llama-2-7b.Q4_K_M.gguf